web/backend/requirements.txt
flask
flask-cors
torch
diffusers
pillow
peft
transformers
accelerate